The SetEQ/GetEQ actually use i2

+++ noson/src/sonosplayer.cpp
@@ -336,7 +336,7 @@ bool Player::SetMute(const std::string& uuid, uint8_t value)
   return false;
 }
 
-bool Player::GetNightmode(const std::string &uuid, uint8_t *value)
+bool Player::GetNightmode(const std::string &uuid, int16_t *value)
 {
   for (RCTable::const_iterator it = m_RCTable.begin(); it != m_RCTable.end(); ++it)
   {
@@ -346,7 +346,7 @@ bool Player::GetNightmode(const std::string &uuid, uint8_t *value)
   return false;
 }
 
-bool Player::SetNightmode(const std::string &uuid, uint8_t value)
+bool Player::SetNightmode(const std::string &uuid, int16_t value)
 {
   for (RCTable::const_iterator it = m_RCTable.begin(); it != m_RCTable.end(); ++it)
   {
@@ -376,7 +376,7 @@ bool Player::SetLoudness(const std::string &uuid, uint8_t value)
   return false;
 }
 
-bool Player::GetSubGain(const std::string &uuid, int8_t *value)
+bool Player::GetSubGain(const std::string &uuid, int16_t *value)
 {
   for (RCTable::const_iterator it = m_RCTable.begin(); it != m_RCTable.end(); ++it)
   {
@@ -386,7 +386,7 @@ bool Player::GetSubGain(const std::string &uuid, int8_t *value)
   return false;
 }
 
-bool Player::SetSubGain(const std::string &uuid, int8_t value)
+bool Player::SetSubGain(const std::string &uuid, int16_t value)
 {
   for (RCTable::const_iterator it = m_RCTable.begin(); it != m_RCTable.end(); ++it)
   {
