+++ noson/src/sonosplayer.h
@@ -95,12 +95,12 @@ namespace NSROOT
     bool GetMute(const std::string& uuid, uint8_t* value);
     bool SetMute(const std::string& uuid, uint8_t value);
 
-    bool GetNightmode(const std::string& uuid, uint8_t* value);
-    bool SetNightmode(const std::string& uuid, uint8_t value);
+    bool GetNightmode(const std::string& uuid, int16_t* value);
+    bool SetNightmode(const std::string& uuid, int16_t value);
     bool GetLoudness(const std::string& uuid, uint8_t* value);
     bool SetLoudness(const std::string& uuid, uint8_t value);
-    bool GetSubGain(const std::string& uuid, int8_t* value);
-    bool SetSubGain(const std::string& uuid, int8_t value);
+    bool GetSubGain(const std::string& uuid, int16_t* value);
+    bool SetSubGain(const std::string& uuid, int16_t value);
 
     bool GetBass(const std::string& uuid, int8_t* value);
     bool SetBass(const std::string& uuid, int8_t value);
